requirements.txt
google-cloud-translate
flask
gunicorn
